Remove warning or error float index.

+++ fluidsim/operators/CySources/operators_cy.pyx
@@ -280,7 +280,7 @@ cdef class GridPseudoSpectral2D(Operators):
                             (self.deltaky*self.ny_seq)**2)/2
 
     def where_is_wavenumber(self, kx_approx, ky_approx):
-        ikx_seq = np.round(kx_approx/self.deltakh)
+        ikx_seq = int(np.round(kx_approx/self.deltakh))
 
         if ikx_seq >= self.nkx_seq:
             raise ValueError('not good :-) ikx_seq >= self.nkx_seq')
@@ -300,7 +300,7 @@ cdef class GridPseudoSpectral2D(Operators):
 
             ikx_loc = ikx_seq - self.iKxloc_start_rank[rank_k]
 
-        iky_loc = np.round(ky_approx/self.deltaky)
+        iky_loc = int(np.round(ky_approx/self.deltaky))
         if iky_loc < 0:
             iky_loc = self.nky_loc+iky_loc
